backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.1
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
supabase==1.0.3
email-validator==2.0.0.post2 